Remove unuseful file boundary
